refactor: consolidate optimizer
