+++ src/node.ts
@@ -143,6 +143,6 @@ export async function run(args: {
 
 export { createNodeFileSystem } from "./vfs/createNodeFileSystem";
 
-export { parseAndEvalExpression } from "./interpreter";
+export { parseAndEvalExpression } from "./optimizer/interpreter";
 
 export { showValue } from "./types/types";
