+++ src/optimizer/constEval.ts
@@ -1,15 +1,18 @@
-import { CompilerContext } from "./context";
+import { CompilerContext } from "../context";
 import {
     AstBinaryOperation,
     AstExpression,
     AstUnaryOperation,
     isLiteral,
     AstLiteral,
-} from "./grammar/ast";
-import { TactConstEvalError, throwInternalCompilerError } from "./error/errors";
-import { AstUtil } from "./optimizer/util";
-import { ExpressionTransformer } from "./optimizer/types";
-import { StandardOptimizer } from "./optimizer/standardOptimizer";
+} from "../grammar/ast";
+import {
+    TactConstEvalError,
+    throwInternalCompilerError,
+} from "../error/errors";
+import { AstUtil } from "./util";
+import { ExpressionTransformer } from "./types";
+import { StandardOptimizer } from "./standardOptimizer";
 import {
     Interpreter,
     InterpreterConfig,
@@ -18,7 +21,7 @@ import {
     evalUnaryOp,
     throwNonFatalErrorConstEval,
 } from "./interpreter";
-import { SrcInfo } from "./grammar";
+import { SrcInfo } from "../grammar";
 
 // Utility Exception class to interrupt the execution
 // of functions that cannot evaluate a tree fully into a value.
